reverted file structure metric
