+++ autoop/core/ml/pipeline.py
@@ -5,7 +5,7 @@ from autoop.core.ml.artifact import Artifact
 from autoop.core.ml.dataset import Dataset
 from autoop.core.ml.model import Model
 from autoop.core.ml.feature import Feature
-from autoop.core.ml.metric.metric import Metric
+from autoop.core.ml.metric import Metric
 from autoop.functional.preprocessing import preprocess_features
 import numpy as np
