+++ autoop/tests/test_pipeline.py
@@ -7,7 +7,7 @@ from autoop.core.ml.dataset import Dataset
 from autoop.core.ml.feature import Feature
 from autoop.functional.feature import detect_feature_types
 from autoop.core.ml.model.regression import MultipleLinearRegression
-from autoop.core.ml.metric.metric import MeanSquaredError
+from autoop.core.ml.metric import MeanSquaredError
 
 class TestPipeline(unittest.TestCase):
